requirements.txt
torch>=2.1
torchvision>=0.16
sentence-transformers>=2.7
nltk>=3.8
numpy>=1.26,<2.0
pillow>=10.0
tqdm>=4.66
matplotlib>=3.8
pyyaml>=6.0
